backend/requirements.txt
aiofiles==25.1.0
aiogram==3.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asyncpg==0.31.0
attrs==25.4.0
billiard==4.2.4
celery==5.6.1
certifi==2025.11.12
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.8.0
fastar==0.8.0
frozenlist==1.8.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.2
magic-filter==1.0.12
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
orjson==3.11.5
packaging==25.0
prompt_toolkit==3.0.52
propcache==0.4.1
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
redis==7.1.0
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.48.0
shellingham==1.5.4
six==1.17.0
sqladmin
SQLAlchemy==2.0.45
starlette==0.50.0
Telethon==1.42.0
tenacity==9.1.2
typer==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.2
uvicorn==0.40.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
phonenumbers
passlib[argon2]
python-jose[cryptography]
fastapi-mail
aiosmtplib
structlog